requirements.txt
fastapi
uvicorn
PyPDF2
nltk
sentence-transformers
faiss-cpu
google-generativeai
python-dotenv
databases
sqlalchemy
asyncpg
requests
psycopg2-binary